requirements.txt
numpy
Pillow
matplotlib
scipy
pandas